Fixed snapping to grid. Find the closest available gridpoint and place the shape on that point. Added 'display grid points' function to drawGrid().
In order to determine the closest available grid point, I built a (bit
of a) brute-force solution. I get all available grid points, determine
the distance towards the gridpoint that was used to drop the shape on,
and then loop through the available grid points (ordered by distance to
droppoint), and test if this point is a valid snappoint. When no valid
snappoint is found, for now I just display a message. Should prevent
shape from being placed at all in the future.

I have added a toggle to displayGrid() so you can make the available /
occupied gridpoints visible. This tends to be very slow, but useful for
testing purposes.